src/requirements.txt
django==1.9
#djangorestframework==3.7.7
djangorestframework==3.3
django-filter==1.1.0
django-crontab==0.7.1
requests==2.18.4
mysqlclient
xlsxwriter